requirements.txt
DateTime==4.3
h5py==2.8.0
Keras==2.1.6
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keras-vis==0.4.1
matplotlib==2.2.2
numpy==1.19.5
opencv-python==3.4.3.18
openpyxl==2.5.6
packaging==17.1
pandas==0.23.4
pathlib==1.0.1
pathlib2==2.3.2
pickleshare==0.7.4
Pillow==5.2.0
python-utils==2.3.0
resnet==0.1
scikit-image==0.14.5
scikit-learn==0.20.0
Theano==1.0.5
toolz==0.9.0
torch==1.8.0
torchsummary==1.5.1
torchvision==0.9.0
tqdm==4.26.0
tensorflow==1.14.0
tensorboard==1.14.0
tensorboardX==1.4
widgetsnbextension==3.4.1
xgboost==1.4.2 
pydicom==2.2.2
imgaug==0.4.0
flask
flask_login
flask_migrate
flask_wtf
flask_sqlalchemy==2.*
python-decouple
sqlalchemy==1.3.18
pyjwt==2.1.0
config
